test: working on AppInformationServiceTests
